+++ passes/R005/testdata/src/a/comment_ignore.go
@@ -1,0 +1,18 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
+)
+
+func fcommentignore() {
+	var d schema.ResourceData
+	
+	_ = d.HasChange("attr1") || d.HasChange("attr2") //lintignore:R005
+
+	//lintignore:R005
+	_ = d.HasChange("attr1") || d.HasChange("attr2")
+
+	//lintignore:R005
+	_ = d.HasChange("attr1") ||
+		d.HasChange("attr2")
+}